add wtf alias, separate checks and report
